Implement parse disambiguation
If a new tree has the same semantic as a previously constructed (and
thus cheaper) tree, then reject the new tree.

If a new tree has a unique semantic but the same display text as a
previously constructed tree, then save the new semantic to the existing
tree's 'disambiguation' array and reject the new tree.