Only allow 'repositories-visibility()' for 1-sg
Only the user has access to his or her private GitHub repositories, so
the semantic for searching public or private repositories is limited to
first-person singular (i.e., "my repos"). This is accomplished by
creating a subcategory of 'repositories' with the rules for
'repositories-visibility()' which can only be used with the rules for
repos created by the user. This has no noticeable impact on performance.

An alternative method is to have the 'repositories-visibility()'
semantic require 'repositories-created(me)' always be in the same RHS
and discard the rule otherwise. This impacts performance by checking
every RHS semantic for if this requirement exists.