Save `text` as array only if needs conjugation
Reduces calls to 'conjugateTextArray()', which traverses the text array
and checks if each item is an Object, by 92%. Despite this enormous
reduction, the overall test is now about 200 ms longer. The cause for
this change is unknown. The `text` property of insertion rules must now
be checked if it is a String or Array, but this is fewer checks than
what occurred in 'conjugateTextArray()'.