Only compare trees with different initial rules
When checking for ambiguity, which compares `nextNodes` of trees with
identical terminal symbols, only compare trees with different root
nonterminal rules. This ensures every instance of ambiguity is found at
the symbol at which the two trees demonstrating ambiguity diverge.

This optimization allows for skipping the ambiguity check of
nonterminal symbols with only one non-edit rule.