Only build trees for distinct pairs of rules
When searching for ambiguity produced by a nonterminal symbol, limit to
distinct pairs of rules produced by the initial nonterminal symbol.
When searching the symbol again to construct the parse trees, only
build paths for the symbol's rules that are in a pair, only print one
instance of ambiguity for each pair, and stop searching after printing
for each pair.

Each instance of ambiguity in the grammar is modeled by a distinct pair
of rules from the a given nonterminal symbol. The previous
implementation printed the same instance of ambiguity multiple times.