Refactor use of pairs of ambiguous trees
When constructing parse trees, each initial path has an array of
`pathTab` indexes of the paths with which it is ambiguous. The array is
shared with all subsequent paths built from the initial path. When
finding an instance of ambiguity, remove the `pathTab` index from the
other path's array. This stops all other paths in the search (from the
same initial paths) from checking for ambiguity with the removed index
(or stops path construction if no indexes remain).