Prevent recursive rules in path search
Prevent recursive rules, where the RHS includes the LHS symbol, in path
search. Recursive rules are permitted for the root nonterminal symbol
of a search (where ambiguity can exist).

Any instance of ambiguity demonstrated by a recursive rule (excluding
for the root nonterminal symbol) can be represented with a different
production because all nonterminal symbols with a recursive rule
require a stop case (an error is thrown in grammar construction if
otherwise).