Prevent `break` if found ambiguity and `printAll`
If `printAll` is `true`, every unique instance of ambiguity generated
by a pair of start rules is printed. Previously, if ambiguity is found
between a pair of paths, the search would skip to the next `pathTab`
index. If this skip does not occur and multiple instances of ambiguity
are found with the same pair of `pathTab` indexes during the same
search, then it is due to ambiguity within that other start rule and it
would be reasonable to `break`. Based on the ordering of the start
rules in the grammar, however, these same ambiguous path pairs could be
found at different times and thus cannot stopped in the same way.
Hence, to avoid inconsistency due to ordering of the start rules in the
grammar, always show the ambiguity and do not break.