Use linked lists for paths' RHS symbols
Use linked lists instead of arrays to maintain each paths' RHS symbols.
This eliminates need to clones the array for each new path, resulting
in a ~15% speed improvement. The ambiguity check is now equal in
performance to the previous implementation (search paths a second time
to construct trees), and 15% faster when `printAll` is `true`.