Sort symbols' rules in grammar by increasing cost
When multiple insertions exist for the same symbol with the same
non-inserted RHS symbol, the `ruleProps` for those insertions are
stored in an array for a single action in the state table, and hence a
single node in the parse forest.

When calculating the minimum cost of a path for the A* heuristic, the
first `ruleProps` in an array of `ruleProps` is used as the minimum
cost for that node. Previously, before symbols' rules were properly
sorted, this was not always the minimum cost, causing some output to
not be in increasing order.