Add support for deletions to `Parser`
Each term in the list of deletables can be deleted when recognized in
input, and each term is only 1 token in length. A sequence of multiple
tokens can also be deleted.

For the token(s) that immediately follow the deletable tokens, nodes
are created for each terminal rule to the token as normal, as well as a
second set of nodes for the same terminal rules with a larger `size`
value for the larger token span and a cost penalty of 1 per deleted
token.