Refactor `treeIsUnique()` in 'forestSearch.js'
Checking for duplicate text in a separate loop is faster than a single
loop because there are ~200x a many semantic duplicates as display text
duplicates. This decreases display text comparisons by 75% by avoiding
comparisons on trees that eventually fail for duplicate semantics. This
increases the speed of the test suite by ~3%.