Add `getFormattedStackFrame()` to 'util.js'
Refactor `util.getPathAndLineNumber()` and
`util.getModuleCallerPathAndLineNumber()` by abstracting the operations
which capture and filter the stack trace and format the chosen stack
frame.

`getFormattedStackFrame()` passes a structured stack trace as an
`Array` of `CallSite` objects, without frames for native Node functions
and this file, to the provided `getFrameFunc(stack)`, which returns a
single frame from the `Array`. It returns the file path and line number
of the returned frame in the format "path:line-number".