Alter handling of complex types in `format()`
Do not indent successive lines if the first argument is an instance of
`Object`. Previously, this only applied to plain `Object`s.

Format any object with a multi-lined string representation on a
separate line. Previously, this only applied to plain `Object`s.

Format all objects with single-line representations on the same line as
the previous argument if the previous argument has a primitive type.
Previously, this occurred whenever the previous argument was a plain
`Object`.