rules: Add nominative rules with negation
Add rules for "repos I do not like" and "repos I have not liked".

Amongst many changes, this requires changing '[do]' to a verb for "I do
not" vs. "{user} does not". In rules which already used the symbol,
however, the grammar now requires conjugation of '[do]' where we
already know the person-number (e.g., "people who [do]..."). This is an
instance, among many in the grammar, where we could create a duplicate
set of terminal rules for an existing verb for use in rules where we
already know the conjugation: the objective and passive cases.
Conjugation is only required during runtime for the nominative case. We
will wait until the grammar is more complex to explore creating
additional versions of existing verbs that do not require conjugation.