cli: Abstract spawning of child processes
`spawnChildProcessAndRedirectStdio()` spawns a new process with
`command` and the command line arguments in `args`. This asynchronous
process avoids blocking the event loop.

Disables CLI stdio while the child is processing, except for `^C`
(`SIGINT`) to terminate the child. Restores the CLI stdio when the
child exits or terminates. This enables terminating the child process
without terminating the CLI itself.

Executes the optional callback after the child process exits and before
returning control to the CLI.