benchmark: Add option to parse as child processes
Integrates 'parseForkable.js' and adds the option
`--use-child-processes`, which serially parses each test query in a
newly spawned process. This avoids the impact of a process's cache on
the benchmark measurements, yielding more consistent measurements.
However, performance still varies slightly due to system cache and other
concurrent processes on the system.

This option is significantly slower because ~250 ms are required to
instantiate the `StateTable` in each new process. (This time is
excluded from the benchmark measurement.)