benchmark: Remove `--parallel` option
As noted, the  `--parallel` option never provided accurate measurements
due to the impact the parallel processes had on system performance.

The preceding `--load-parallel` option was slightly accurate, though
still hindered by the system performance impact of ~350 waiting
processes. The 3x speed improvement over the `--serial-child-processes`
option is immaterial as it is still 40 seconds. Hence, the option will
not be restored.