old: Add 'splitMultiTokenTerminalSymbols.js'
This saves the attempted improved implementation of terminal rules.
This implementation is described at length in the file. After much
work, thought, trials, and frustration, this work is being temporarily
halted and will be resumed later. It is too demoralizing to bang one's
head against the wall for so long. Continued processing in the back of
the mind, accompanied with progress on other areas of the system, will
make the implementation easier to accomplish in the future.