forestSearch: Add `appendSemantic()`
`appendSemantic()` abstracts the operation previously in `createPath()`
to append a new non terminal rule's semantics to the previous path's
semantic list.

- If `newSemantic` exists and is reduced (i.e., RHS), then it is merged
with the previous semantic if it too is a reduced.

- If `insertedSemantic` exists, then it is a RHS semantic and
`newSemantic` also exists and is a LHS semantic.

- Fails when appending a LHS semantic and the previous semantic is
completely reduced (i.e., RHS), is identical to `newSemantic`, and
multiple instances of the semantic are forbidden, and then returns `-1`.

- Fails when attempting to merge RHS semantics which would illegal
(e.g., are duplicates), and then returns `-1`.