ambig: Check for ambiguity from insertion rules
A pair of paths is ambiguous when their rightmost symbols are identical
and have identical semantic trees and/or identical display text.

Extends `createPath()` to save insertion text of the second symbol of a
binary rule in `nextSyms`, and to conjugate that text after completing
the corresponding branch.

Adds `conjugateText()` and `conjugateTextArray()`, which conjugates text
of normal and edit rules using paths' grammatical properties.

Adds `textsEqual()`, which performs a comparison between two rules'
text to determine if they are equivalent.

Adds `nextTextsEqual()`, which compares two paths' `nextSyms` linked
lists of symbols and yet-to-be-conjugated text, ignoring the instances
of symbols, to determine if the text values are equivalent.

Amends `nextSymsEqual()` to ignore yet-to-be-conjugated text within
`nextSyms`.

Amends `linkedListToGraph()` to include insertions in graphs.