semantic: Refactor `intersect` semantic comparison
Amends `semantic.reduce()` to compare the LHS semantic to the actual
`intersect()` semantic instead of comparing the semantics' `name`
properties. If not currently building the grammar, then loads the
`intersect()` semantic from the grammar file, which works for
comparisons because of 'initSemantics.js'. Else saves the semantic upon
its creation.