benchmark: Listen for 'exit' from child processes
Amends `childProcessesBenchmark` to listen for the 'exit' event instead
of the 'close' event when waiting for child processes to exit.

The 'exit' event is emitted when the child process ends, while the
'close' event is emitted when the stdio streams of a child process have
all terminated. This change has no effect on this program; however,
'close' should only be used in the special cases it applies.