edit-rules: Detect attempted ambiguous rules
Throws an exception when attempting to add an insertion rule that
creates ambiguity. I.e., a rule with identical RHS symbols and
identical display text and/or semantics as an other rule.

Previously, when attempting to add an insertion with identical RHS
symbols and semantics, `ruleIsUnique()` chose the rule with the lowest
cost. This change forces the user to modify the offending grammar rules.