edit-rules: Detect ambiguous insertions
Throws an exception when attempting to add an insertion that creates
ambiguity. I.e., an insertion with identical display text and/or
semantics as an other insertion.

Previously, when attempting to add an insertion, `insertionIsUnique()`
chose the rule with the lowest cost. This change forces the user to
modify the offending grammar rules.