test: Add optional check of all semantic results
Adds support for optionally defining `test.suggestions`, an array of
all of the test's expected semantic results for its query.

If the command line option `--all-semantics` is provided, compares a
test's semantic results to the test's expected semantics. Checks for
missing and new semantics, ignoring order, and prints the differences.

Adds `test.suggestions` for the majority of existing tests.