test: Add `addTest`
`addTest` is a command line program that parses a provided query and
finds the 50-best parse trees, creates a new test from the parse trees,
and then saves the new test to the test suite.

`addTest` aborts if a test for the provided already exists, else
outputs the new test for the user to confirm before saving it.