test: Add `updateTests`
`updateTests` is a command line program that parses and finds the
50-best parse trees of the existing tests queries in the test suite,
creates new tests with expected values derived from the parse trees,
and then overwrites the existing tests in the test suite. If the new
test differs from the existing test, outputs the new test for the user
to confirm before overwriting.