rules: Stash '/old/CategorySansAmbig/'
Saves an unused implementation of `Category`, which removes instances
of semantic ambiguity produced by queries of the following forms:
- "repos I like"
- "repos that I like"
- "repos I like Danny likes"
- "repos I like and Danny likes"
- "repos that I like and Danny likes"
- "repos that I like that Danny likes"
- "repos that I like and that Danny likes"
- "repos I like and that Danny likes"
- "repos that I like Danny likes"

While this implementation can significantly reduce the number of
semantically identical trees that are generated and then discarded, the
required rules yield an increased total number of created paths. More
importantly, the greater number of rules can significantly increases
the time required for `Parser.prototype.parse()` (i.e., the parse
forests are more complex). As a result, this implementation is slower.