grammar: Refactor verbs that accept varying tenses
Previously, if a verb was accepted in multiple forms (e.g., "repos I
like/liked") the property `alwaysAcceptPast` was used to create a new
verb. For other rules that used the same terminal symbols for which
multiple forms could not be accepted (e.g., "repos I do not like"), a
second verb with the same symbols was created.

This change removes `alwaysAcceptPast` for verbs and adds the
nonterminal rule property `acceptedTense` and terminal rule property
`tense`. If a nonterminal rule defines `acceptedTense`, and the rule
produces a terminal symbol with a matching `tense` property, then that
terminal symbol (i.e., verb form) is accepted (i.e., not corrected)
when seen in input, but not forced to that form when otherwise.