grammar,parse: Significantly extend insertions
Rules where the inserted segment is in the right (second) position of a
binary rule and includes the `[1-sg]` symbol in the insertion, only
insert at the end of an input query.

	1. This reduces the number of semantically ambiguous suggestions
	otherwise created by inserting the same query segment (e.g., "liked by
	me") at multiple positions in the input query.

	2. This limits the number of such insertions at the end the input
	query to one.

  3. Most importantly, this prioritizes suggestions that are easier for
	the user to read by only inserting at the end. In contrast,
	arbitrarily positioning edits anywhere in the query disorients the
	user as they must discern the suggestion's differences from input.

The `Parser` inputs the `<blank>` symbol at the end of parsing the
input query to only allow these binary insertion rules with `<blank>`
to match at the end of the query.

The `<blank>` symbol also enables deletions (with costs) for input
tokens that precede it (i.e., the last token(s) of input).

In `pfsearch`, the insertion text is handled just as normal insertions
because the second branch for `<blank>` is not traversed, but also
count any deletion costs `<blank>` may absorb.

Updates expected semantics of existing tests to reflect these changes.