Parser: Enable deletions at end of input query
Extends `Parser` to detect deletables at the end of the query.

After checking for continuous spans of deletables that precede a
terminal symbol, checks if all tokens that follow that terminal symbol
(through the end of the input query) are deletable. If so, creates new
nodes for the span beginning with the terminal symbol and ending with
the deletables that extend to the end of input, and for each distinct
span of preceding deletables already found, if any, ending with the
deletables that extend to the end of input. Hence, if the input query
ends with deletables, the deletion cost is added to the preceding
terminal symbol, and for each node created for that terminal symbol for
any preceding deletables.

Previously, this only occurred for insertion rules with the `<blank>`
symbol by adding the deletion cost to `<blank>`. Now, only adds
deletion costs to the last terminal symbol (not `<blank>`), and
`pfsearch` no longer checks for deletion costs associated with
`<blank>`.

Adds tests and updates existing tests to reflect this change.