symbol: Add multi-symbol text substitution
Extends `Symbol` to support multi-symbol text substitution; e.g.,
"about","which" → "that". Adds `text` and `costPenalty` properties to
nonterminal rules, which replace the terminal symbols the rules' RHS
produce when seen in input. This is necessary to restrict terminal
symbols to individual tokens.

Adds `Symbol.prototype.addBlankSet()`, which adds terminal to a
`Symbol` that do not produce display text when seen in input, but
rather uses text defined on a parent nonterminal rule.