symbol: Refactor multi-symbol substitutions.
Instead of manually creating rules for multi-symbol substitutions,
substitutions are now defined using a regex-style string.

Adds `splitMutliTokenTerminal()`, which tokenizes a terminal symbol by
spaces (separate terms indexes) and the `|` symbol (acceptable terms
for that single index). Creates an array of `Symbol`s that produces a
tree of single-token terminal rules for the tokens, for use as a
nonterminal rule's `RHS`.

Removes `Symbol.prototype.addBlankSet()` and `g.newBlankSet()`.

Limits `substitutions` for all `Symbol` methods to be defined as
strings or objects with properties `symbol` (string) and `costPenalty`
(number).