symbol,edit-rules: Add `restrictInsertion`
Adds `restrictInsertion` parameter to terminal symbols, which specifies
insertions that include the terminal rule, and where the inserted
segment is in the right (second) position of a binary rule, can only
occur at the end of an input query. Previously, this was implemented by
hard-coding the terminal rules in `createEditRules`.