semantic: Enable semantics to require others
Extends semantics to enable semantic functions to require other
semantics in the same semantic tree and within the same instances of
`intersect()`, else the semantic tree is rejected.

This is used for queries such as, "my private repos", where
`repositories-visibility(private)` requires `repositories-created(me)`,
because only the user has access to the data for such queries (i.e.,
private repos).

This check occurs in `semantic.reduce()`, when the LHS semantic is
`intersect()` (and the RHS is complete). This must occur at
`intersect()`, else semantics such as the following would be rejected
(because the required semantic is not within `not()`):
`not(repositories-visibility(public)),repositories-created(me)`.

Adds `isMissingRequiredSemantic()`, which checks if `semanticArray`
contains a semantic that `requires` another semantic not within this
semantic tree, stopping before instances of the `intersect()` semantic.

Adds `hasSemantic()`, which checks  if `semanticArray` contains
`semanticNode`, stopping before instances of the `intersect()` and
`not()` semantics.