old: Archive `DHeap`
Archives `DHeap`, a generalization of `BinaryHeap` in which the nodes
have `d` children instead of 2.

To use a `d`-ary heap with a a shortest path graph algorithm, where `n`
is the number of vertices in the graph and `m` is the number of edges,
then `d = m/n` to balance the total times for these two types of
operations against each other. Using this calculation with the parse
forest graphs which `Parser` produces from this grammar, `d` almost
always rounds to 2. Hence, `pfsearch` uses `BinaryHeap` (i.e., where
`d` is 2) instead of `DHeap` to avoid accommodating for scarcely
variable arity.