Parser: Delete unrecognized input tokens
Adds deletion of unrecognized tokens in the input query to
`Parser.prototype.deletablesLookup()` in `matchTerminalRules`. This
occurs when an input token is not a recognized lexical token in the
query language or the entities, nor is a legal integer symbol (i.e.,
within the symbols' specified bounds).

Adds tests and updates expected values of existing tests to reflect
this change.