StateTable: Add `calcMinShiftsToAccept`
Adds `calcMinShiftsToAccept`, which calculates the minimum number of
additional shifts needed at each shift and reduction to reach the state
table's accepting state. This enables discarding nodes in `Parser` we
know fail because too few input tokens remain. Also helps identify when
a sequence of terminal nodes is impossible and requires a deletion.