parse: Deprecate `calcMinShiftsToAccept`
Deprecates `calcMinShiftsToAccept` because `minShiftsToAccept`, the
minimum number of additional shifts at each shift and reduction of every
state needed to reach the state table's accepting state, is not useful.
These values are almost always 0-2, which is rarely helpful because
queries are rarely that short. This implementation also adds additional
overhead to check these values for every shift and reduction.

In addition, these values are inapplicable to detecting which input
tokens recognized by the query language we can delete in order to avoid
failing to reach the start node. An enormously simpler and correct
solution is currently being developed.