parse: Add `initEntities`
Adds `initEntities`, which initializes the entities in the grammar for
parsing by replacing multiple instances of the same entity in the map
of tokens to entities with references to the same object. I.e., this is
only applicate to multi-token entities where there are multiple maps to
the same entity. This enables equality checks by object reference
instead of checking their `id` properties.

Integrates `initEntities` with `buildStateTable`.

Replaces the entity equality check in `matchTerminalRules` by `id` with
a comparison of object reference.

This change atomically improves performance and slightly reduces memory
usage.