test: Add support for tagging tests
Extends `test` to support a `tags` property on tests in the test suite
for categorization. Invoking `test <tag>` only runs tests with the
provided tag. `test` can be invoked with multiple tags (as command line
arguments) to run tests with any of the provided tags. This extension
enables focused testing, such as running only those tests that fail to
reach the start symbol on the initial parse.