Parser: Prevent ambiguity in multi-token entities
Extends `Parser.prototype.mergeEntityMatches()` to avoid adding
duplicates to `matchedEntities`. If a new merged entity match is
identical to an existing one, keeps the cheapest. Deletions of entity
tokens cause this, which occur when an entity token is a
grammar-defined deletable.

Adds `getMatchedEntityIndex()`, which gets the index of where to add
the entity with the provided parameters in `matchedEntities` to avoid
duplicates in `Parser.prototype.mergeEntityMatches()`.