pfsearch,test: Always return path count with trees
Amends `pfsearch` to always include `pathCount`, the number of paths
created in a parse forest search, as a property of the returned array
of parse trees. Removes the `pfsearch` option
`options.returnPathCount`, which previously specified returning an
object with properties `trees` and `pathCount`.

Updates `test` accordingly, which uses `pathCount` to benchmark the
number of paths created from parsing the entire test suite.