Parser: Re-parse after producing no legal trees
After `pfsearch` fails to generate legal parse trees (due to
contradictory, illegal semantics), resets the cost heuristics, marks
all tokens as deletable (except those already added to
`this.deletions`) and adds new nodes with those deletions, just as with
parses that fail to reach the start node.

Adds `Parser.prototype.resetMinCosts()`, which resets the `minCost`
property of all nonterminal subnodes to `undefined`. Most of the nodes
will be reused in the second parse, and therefore require their
`minCost` value be set to `undefined` to enable `calcHeuristicCosts` to
re-calculate the heuristic costs of the new parse forest.

Updates the expected values of existing tests (with the tag 'no-trees',
for previously failing to generate parse trees) to reflect this change.