rules: Add anaphora
Adds anaphoric rules, such as "people who follow {user} and his
followers", where the semantic for the antecedent "{user}" is copied in
the place of the anaphor, "his".

Extends `semantic.reduce()` to perform the above operation. Requires
adding a special `anaphoric` property the LHS semantic for the anaphor
(which above is `followers()`).

Amends `initSemantics` to use different semantic comparison functions
for equality checks, to check for the `anaphoric` property.