rules: Remove previous anaphora implementation
Remove the previous anaphora implementation, including the semantic
with the special `anaphoric` property, the expanded `initSemantics`
modules to check for the property, the new rule, and the modified
existing rule modules to enable the rules.

A far simpler, more powerful, and entirely different approach has been
discovered and will replace this implementation.