grammar,parse: Refactor anaphora resolution
Adds `anaphoraPersonNumber` property to nonterminal rules, which is the
person-number for anaphoric rules with which to match to an antecedent
semantic of the same person-number. E.g., "his/her" refers to semantics
of third-person-singular representations.

Removes anaphoric semantic arguments.

Lowers the cost penalties for first-person-singular terminal rule set
substitutions with anaphoric terms in proportion to the removal of the
cost of anaphoric semantic arguments.

Extends `ruleMissingRHSSemantic` to check for `anaphoraPersonNumber`.

Extends `createRuleProps()`, invoked during `StateTable` generation, to
include `anaphoraPersonNumber` in `ruleProps`.

Amends `pfsearch` to check for `anaphoraPersonNumber` instead of
searching a rule's RHS semantic for anaphoric semantics. This allows
for the removal of `resolveAnaphora()`, which searched for such
semantics. Removes invocation of anaphora resolution from
`reduceSemanticTree()` because `anaphoraPersonNumber` only occurs on
nonterminal rules.

Updates the expected semantics of existing tests (whose only changes
are their order due to floating point precision) to reflect these
changes.