rules: Increase anaphoric substitution costs
Increases the cost of substitutions with anaphoric terms because these
substitutions primarily exist to handle illegal anaphora, and not to
vary suggestions. Edits are to expand on input and handle ill-formed
input, but not to offer alternatives to input queries.

Updates the order of the expected semantics of existing tests to
reflect these changes.