test: Improve default input for `addTest`
Extends `addTest` to improve the default input suggestion for the new
test's top result semantic. If the user provides a top result display
text that differs from the parse's and the input query, then parse that
(corrected) display text and use its top result semantic as the
suggested (default) semantic instead of the original parse's.