grammar: Sort entities in grammar for each token
Extends `grammar.sortGrammar()` to sort entities for each token
alphabetically by display text and then by name tokens (which can
differ). The previous implementation incorrectly did not sort entities.

Sorts by display text first so that parse trees with entity matches to
the same input token and with the same match cost are sorted
alphabetically when output.

Updates the order of expected semantics for existing tests to reflect
this change.