benchmark: Change default `k` value to 7
Changes default `k` value for `pfsearch` from 50 → 7. To properly
benchmark the program, it must be measure as how it will run in
production. This change demonstrates an enormous, fundamental
misunderstanding which has lasted months: `Parser` consumes ~85% of
processing while `pfsearch` consumes ~10%, as opposed to 50% and 40%,
respectively when `k` is 50 (as in `test`). Ergo, the processing cost of
trees which `pfsearch` discards for ambiguity is negligible compared to
the overhead `Parser` endures (in shift and reduce actions) for the
extra rules needed to avoid ambiguity.