Parser: Avoid ambiguity with entity aliases
Extends `Parser.prototype.addMultiTokenEntityNodes()` to check for
matches to the same entity (id) via multiple (ambiguous) names/aliases,
over the same token span, and keep the cheapest match. E.g., a match to
"Alan" can be for either the alias "Alan Kay" or "Alan Curtis", both of
which map to "Alan Kay". Keeps multiple matches to the same entity (id)
via multiple names/aliases over different token spans, just as multiple
matches to the same entity object over different spans are kept, as all
possible spans must be examined to explore possible insertions. It is
likely the overhead from this additional check for this obscure edge
case is more detrimental to performance than letting the `Parser` and
`pfsearch` take on additional load, and allowing for the ambiguity to
be caught at the end of `pfsearch`.

Adds tests for ambiguous matches to the same entity via different
entity objects.

Adds the entity "Alan Kay" with alias "Alan Curtis" to test ambiguous
matches to the same entity via different entity objects.

Updates the semantics of existing tests for the entity id bumps from
this new entity.