grammar: Specify entity name ambiguity potential
Extends `entityCategory` to define `hasAmbiguousMultiTokenAlias` for
entity aliases, which specifies if an entity name is multi-token and
shares a token with another multi-token name for the same entity id.

The property instructs `matchTerminalRules` to check for multiple
matches to the same entity id via different aliases over the same token
span, and thereby avoid ambiguity. E.g., a match to "Alan" can be for
either the alias "Alan Kay" or "Alan Curtis", both of which map to
"Alan Kay".

Only checks multi-token names because only multi-token names are added
in `Parser.prototype.addMultiTokenEntityNodes()`, which does not have
access to the uni-token entity matches to check for ambiguity.

Adds `hasAmbiguousMultiTokenAlias()` to `entityCategory` which checks
for this property.