grammar: Define ambiguous uni-token entity aliases
Extends `entityCategory` with `getAmbiguousUniTokenAliases()`, which
gets the uni-token names of an entity with multiple names/aliases which
are contained within the entity's multi-token names, if any.

Extends entities for which this is true with the property
`ambiguousUniTokenAliases`, which contains the tokens or is `undefined`
otherwise (to avoid invoking `indexOf()` on empty arrays).

These tokens (if any) instruct
`Parser.prototype.addMultiTokenEntityNodes()` in `matchTerminalRules`
to avoid adding a parse node for a single-token entity match via a
token for which a uni-token alias for the same entity was already
added. I.e., avoids multiple matches for the same entity (id and
display) via different names/aliases over the same one-token span.

A single-token match to a multi-token entity name for a token returned
by this function will not be added to the parse forest, but only kept
for merging with adjacent matches to the same entity object (to form a
multi-token match), because the ambiguous uni-token alias will be
cheapest.

E.g., a match to "Iroh" can be for either the alias "Iroh" or "General
Iroh", both of which map to "Iroh". The single-token match to "Iroh"
for the name "General Iroh" must be kept to check for an adjacent match
to "General", but no node will added for the match to only "Iroh" for
this name's entity object.

This property supplements `hasAmbiguousMultiTokenAlias`, which only
checks for possible ambiguous matches with other multi-token names
because `Parser.prototype.addMultiTokenEntityNodes()` does not have
access to the uni-token entity matches to check for ambiguity.