Parser: Integrate `ambiguousUniTokenAliases`
Extends `Parser.prototype.entityLookup()` in `matchTerminalRules` to
check if a multi-token entity match has `ambiguousUniTokenAliases`, to
specify if there exists a uni-token alias for the same entity that is
contained within the match's (multi-token) name. If `true` instructs
`Parser.prototype.addMultiTokenEntityNodes()` to not add a parse node
for this single-token entity match because the uni-token alias (for
this entity) for the same token is added below. I.e., avoid multiple
matches for the same entity (id and display) via different
names/aliases over the same one-token span. Keep this match for merging
with adjacent matches to the same entity name (to form a multi-token
match).

Note: Like `hasAmbiguousMultiTokenAlias`, it is likely the overhead
from this additional check for this obscure edge case is more
detrimental to performance than having `Parser` and `pfsearch` absorb
additional load, and allowing `pfsearch` to catch the ambiguity at the
end.