rules,test: Add entity to test ambiguity
Adds an entity with multiple names to the `{user}` category to test
ambiguity where a multi-token name contains a token that is also a
uni-token name for the same entity. A match to only that token matches
both names, which both map to the same entity (id and display text).

Adds tests for this case and updates the expected semantics of existing
tests for the entity id bump (i.e., different semantic arguments) from
the new entity's addition.