Parser: Refactor start node search
Refactors the start node search at the end of
`Parser.prototype.shiftReduce()`. Previously, the method invoked
`getStartNode()` for the last `vertTabs` index and then the
second-to-last index if no start node was found in the last index.

This change replaces `getStartNode()` with
`Parser.prototype.getStartNode()`, which first checks `zNodeSets`,
which maps directly to the accepting state vertex, if any. If none,
this method then iterates through the second-to-last `vertTabs` index
as before.