Parser: Refactor `Parser.prototype.addNode()`
Refactors `Parser.prototype.addNode()` to require the next state (for
the shift) as an argument. Amends invocations of the method
accordingly. These changes reduces checks (do not need to check if a
shift does not exist for a reduction) and variable references (can
cache the node's symbol index for shift lookups in the parent
functions).

These changes improve performance of the system by 3.2%.