Parser: Refactor `Parser.prototype.addSub()`
Refactors `Parser.prototype.addSub()` where `nodeSets` contains a map
for each nonterminal symbol of node size to node (for that symbol),
instead of an array of nodes of each size for that symbol. This
accelerates node lookup and assignment, however, it requires creating
large arrays as maps for the sizes.

Updates `Parser.prototype.resetMinCosts()` to handle `nodeSets` mapping
to sparse arrays.