Parser: Revert `Parser.prototype.addSub()`
Reverts `Parser.prototype.addSub()` implementation to store nodes for
the same symbol at the current index in an array instead of a node size
map. Most often there are only 1 (50%) or 2 (20%) nodes per symbol per
index, so the cost of constructing the array map is greater than the
time saved by removing iteration.