Parser,pfsearch: Assign `minCost` to nodes
Extends `calcHeuristicCosts` to assign `minCost` to nodes in addition
to sub-nodes. This reduces the number of recursive calls (to
`calcMinCost()`) by 50% because previously there were duplicate calls
to the same sets of subnodes.

This change increases the system's overall speed by 1.8%. Likely,
despite reducing recursive calls by 50%, because of the overhead for the
additional object properties. This reduces `calcHeuristicCosts` from
4.4% → 3.3% of the system.