pfsearch: Exclude `minCost` from `next` subnodes
Amends `calcHeuristicCosts` to only assign `minCost` to the second node
in a binary rule, not to the second subnode. This is because the first
subnode (in a binary rule) is distinguished by containing the sum of
its node's `minCost` and its `next` node's `minCost`, whereas the
`next` subnode's `minCost` is identical to its node's `minCost`.

Updates `pfsearch` accordingly.